+++ src/hooks/common/useDebounce.js
@@ -1,0 +1,16 @@
+import { useEffect, useState } from 'react';
+
+export default function useDebounce(value, delay = 1000) {
+  const [debounceValue, setDebounceValue] = useState(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setDebounceValue(value);
+    }, delay);
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value, delay]);
+
+  return debounceValue;
+}
